split up tests more
